website/requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap-datepicker-plus==3.0.5
django-cors-headers==3.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.6.0
kombu==5.1.0
prompt-toolkit==3.0.19
Pygments==2.9.0
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
redis==3.5.3
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
django-import-export==2.5.0